Refs #9958: move headers to private
