+++ src/cpp/rtps/builtin/discovery/participant/PDPClient.h
@@ -23,7 +23,8 @@
 
 #include <fastdds/rtps/builtin/discovery/participant/PDP.h>
 #include <fastdds/rtps/messages/RTPSMessageGroup.h>
-#include <fastdds/rtps/builtin/discovery/participant/timedevent/DSClientEvent.h>
+
+#include "./timedevent/DSClientEvent.h"
 
 namespace eprosima {
 namespace fastdds{
