+++ src/cpp/rtps/builtin/discovery/participant/PDPServer2.cpp
@@ -44,7 +44,7 @@
 
 #include "./PDPServer2.hpp"
 #include "./PDPServerListener2.hpp"
-#include "./DServerEvent2.hpp"
+#include "./timedevent/DServerEvent2.hpp"
 #include "../endpoint/EDPServer2.hpp"
 #include "../endpoint/EDPServerListeners2.hpp"
