+++ src/cpp/rtps/builtin/discovery/participant/timedevent/DSClientEvent.cpp
@@ -17,9 +17,6 @@
  *
  */
 
-#include <fastdds/rtps/builtin/discovery/participant/timedevent/DSClientEvent.h>
-#include <fastdds/rtps/builtin/discovery/participant/PDPClient.h>
-
 #include <fastdds/rtps/builtin/data/ParticipantProxyData.h>
 
 #include <fastdds/rtps/resources/ResourceEvent.h>
@@ -28,6 +25,8 @@
 
 #include <fastdds/dds/log/Log.hpp>
 
+#include "./DSClientEvent.h"
+#include "../PDPClient.h"
 
 namespace eprosima {
 namespace fastdds {
